// src/enums/technology.ts
export enum Technology {
	// Level I
	AnimalHusbandry = 'animalHusbandry',
	Agriculture = 'agriculture',
	CodeOfLaws = 'codeOfLaws',
	Currency = 'currency',
	HorsebackRiding = 'horsebackRiding',
	Masonry = 'masonry',
	MetalWorking = 'metalWorking',
	Navigation = 'navigation',
	Navy = 'navy',
	Philosophy = 'philosophy',
	Pottery = 'pottery',
	Writing = 'writing',

	// Level II
	Bureaucracy = 'bureaucracy', //coin
	CivilService = 'civilService', //coin
	Chivalry = 'chivalry',
	Construction = 'construction',
	Democracy = 'democracy',
	Engineering = 'engineering',
	Irrigation = 'irrigation',
	Logistics = 'logistics',
	Mathematics = 'mathematics',
	Monarchy = 'monarchy',
	Mysticism = 'mysticism',
	PrintingPress = 'printingPress',
	Sailing = 'sailing',

	// Level III
	Banking = 'banking',
	Biology = 'biology',
	Communism = 'communism',
	Ecology = 'ecology',
	Education = 'education',
	Gunpowder = 'gunpowder',
	Metallurgy = 'metallurgy',
	MilitaryScience = 'militaryScience',
	Railroad = 'railroad', //coin
	SteamPower = 'steamPower',
	Theology = 'theology',

	// Level IV
	AtomicTheory = 'atomicTheory',
	Ballistics = 'ballistics',
	Combustion = 'combustion',
	Computers = 'computers', //coin
	Flight = 'flight',
	MassMedia = 'massMedia',
	Plastics = 'plastics',
	ReplaceableParts = 'replaceableParts',

	// Level V
	SpaceFlight = 'spaceFlight',
}